move external_tools
