+++ sharkfuser/include/fusilli.h
@@ -15,8 +15,8 @@
 
 // Support
 #include "fusilli/asm_emitter.h"
-#include "fusilli/external_tools.h"
 #include "fusilli/extras.h"
+#include "fusilli/support/external_tools.h"
 #include "fusilli/support/logging.h"
 #include "fusilli/types.h"
